+++ diwiseweb/src/components/Map/Map.js
@@ -3,6 +3,7 @@ import MapContext from "./MapContext";
 import * as ol from "ol";
 import Overlay from 'ol/Overlay';
 import "./Map.css";
+import { Popup, DevicePopupContent, FeaturePopupContent } from "./Popup";
 
 const Map = ({ children, zoom, center }) => {
 	const mapRef = useRef();
@@ -69,15 +70,6 @@ const Map = ({ children, zoom, center }) => {
 	)
 }
 
-const Popup = ({ parent, popupContent }) => {
-	return (
-	  <div id="popup" ref={parent} className="ol-popup">
-		<a href="#" id="popup-closer" className="ol-popup-closer"></a>
-			<div id="popup-content">{ popupContent }</div>
-	  </div>
-	);
-};
-
 const getPopupContent = (feature) => {
 	if (feature.markerType === "device") {
 		return <DevicePopupContent feature={feature} />
@@ -87,41 +79,6 @@ const getPopupContent = (feature) => {
 	}
 
 	return <></>;
-};
-
-
-const DevicePopupContent = ({ feature }) => {
-	const d = feature["data"];
-	return (
-		<>
-			<div><strong>devEUI:</strong>{d.devEUI}</div>
-			<div><strong>deviceID:</strong>{d.deviceID}</div>
-			<div><strong>Namn:</strong>{ d.name }</div>
-			<div><strong>Beskrivning:</strong>{ d.description }</div>
-			<div><strong>Senast:</strong>{ d.lastObserved }</div>
-		</>
-	);
-};
-
-const FeaturePopupContent = ({ feature }) => { 
-	const f = feature["data"];
-	return (
-		<>
-			<div><strong>Id:</strong>{f.id}</div>
-			<div><strong>Type:</strong>{f.type}</div>
-			<div><strong>SubType:</strong>{f.subtype}</div>
-			{f.type === "counter" ? <CounterPopupContent counter={ f.counter } />:<></>}
-		</>
-	);
-};
-
-const CounterPopupContent = ({ counter }) => {
-	return (
-		<>
-			<div><strong>Antal:</strong>{counter.count}</div>
-			<div><strong>Status:</strong>{counter.state ? "På":"Av"}</div>
-		</>
-	);	
 };
 
 export default Map;
